Added caches option to be able to flush data from identified caches

+++ Command/CacheFlushAllCommand.php
@@ -12,6 +12,7 @@
 namespace Sonata\CacheBundle\Command;
 
 use Symfony\Component\Console\Input\InputInterface;
+use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Output\Output;
 
@@ -24,6 +25,8 @@ class CacheFlushAllCommand extends BaseCacheCommand
     {
         $this->setName('sonata:cache:flush-all');
         $this->setDescription('Flush all information set in cache managers');
+
+        $this->addOption('cache', null, InputOption::VALUE_REQUIRED | InputOption::VALUE_IS_ARRAY, 'Flush elements stored in given cache');
     }
 
     /**
@@ -34,6 +37,10 @@ class CacheFlushAllCommand extends BaseCacheCommand
         $output->writeln('<info>clearing cache information</info>');
 
         foreach ($this->getManager()->getCacheServices() as $name => $cache) {
+            if ($input->getOption('cache') && !in_array($name, $input->getOption('cache'))) {
+                continue;
+            }
+
             $output->write(sprintf(' > %s : starting .... ', $name));
 
             if ($cache->flushAll() === true) {
